example app to root directory
